Solved issues with timing
It really had to do with the fact that I kept on adding to millis
instead of just letting it be equal to the equivalent milliseconds of
the difference between start time and the end time.

So now it seems like it'll keep about the same place for at least a full
minute, which is good.