Added data sharing, fixed minor bug
I added data sharing by making a custom tab bar controller called
InfoSharingTabController that has a field inside it called data that is
used to store the milliseconds that it takes to execute each time.

I also found out that if I just tap the screen, it was adding times of
zero to the list of times. All I had to do to fix this was to add a sort
of "toggle" in the form of a boolean that ensures that whenever I tap,
it has to have been timing in order to also save the data.