Added mechanism for deletion and changing scramble
I added gestures such that a left swipe deletes the last time (with the
added caveat that you should also verify with the UIAlert) and a right
swipe means that the scramble is changed.

Luckily, I'd already figured out all of this insertion stuff for the
table when I made the wellesley fresh app.