Resolved #6: save times from different sessions
I have it solving the times to a single session, FirstSession.

Who knows how I'm going to have it save to multiple sessions.

I guess I will have to make a currentSession -- either an integer or a
string -- that will save the current session for me.

I'm going to have to put that into user defaults as well, so that I can
save the current session from one instance of the app running to the
next.