Resolves #21, Resolves #19, Resolves #20
Issue #21 - add new VC for sessions
Added one to the storyboard. I need to name its segue so that I can call
a programmatic segue on it and open it up for editing.

Issue #19 - Add bottom header + table footer
This just makes it so much cleaner, so I had to add it to
SecondViewController, programmatically. Along with just adding the
bottom header (which was apparently already there), I also recolored the
backgrounds of the headers. Because I was returning an alternative view,
I could no longer rely on titleForHeaderInSection to assign titles to
the headers, so I had to configure thta in viewForHeaderInSection as
well.
The footer view is added in viewWillAppear.

Issue #20 - Add variable to keep track of the current session
This was added to the InfoSharingTBC, so that I didn't have to pull my
hair out when I add support for switching sessions. Because sessions are
already built in, through some methods that I added to InfoSharingTBC,
I think it will be much easier to enable session changing later on in
the app's development.

I normally like to only solve one issue per commit but I'm unfortunately
somewhat ill while programming (not the most prudent choice) and got
carried away while making edits. I have however been testing it along
the way so it should be fine mostly.