Fixed #31
Issue #31 - fix csv file writing
As it happens, there were some more issues at hand than I thought -
mainly that I didn't know that there was an additional issue of getting
the whole thing with the csv file writing done.
The ultimate test of this working would be retrieving the file but I
kind of want to get sharing up and running before I do that.