Fixes #33
Issue #33 - time does not change colors on iPhone 7
I figured out that it was a problem of mach_absolute_time()'s units
being screwy on the iPhone as compared to the computer. On the computer,
it's at least comprable to actual nanoseconds, but on the iPhone, it's
something different, so I ended up making a function for FirstVC that
just gets the elapsed nanoseconds for me.