Working on #32 - sharing bug
This is a bit of a misnomer - it's actually a file bug. I can't get the
file to be saved somewhere that it will be able to retrieve from, and I
can't get the file manager to print out a list of files from the
directory I am supposedly saving my file to.

I'm probably going to branch to a fix branch in order to work on this in
more detail.