Fixes #32: file sharing bug
It...suddenly works? You might understand why I'm skeptical. It all of
a sudden started working after I got rid of a "let url = try [create
directory bits here" and had it print out the current directory.

Go figure.

Not sure if it completely works or anything, but it did send the
appropriate file to my computer, which is a win in my book

An unexpected win, but a win all the same.

Also turns out that mailing works too. I have yet to try any of the
other functions.