Resolves #18: new view controller for editing
In addition to adding a new view controller to the main storyboard, I
also made classes for the respective VCs for editing and for choosing
sessions. I also made sure to give the segues their own identifiers and
identified the classes of each of the VCs in the storyboard, so that it
might be easier in the future to tell them apart.