Resolves #17: add clear times button to edit scene
I actually added it in the form of an action sheet that you access by
clicking on a button in the top right-hand corner called "Options" but I
still added it. I haven't been able to test it because I haven't exactly
put together the whole segue thing but I'm sure I'll get around to that
as soon as I hook it up in SecondViewController, in the same function
where I deal with all of the sharing stuff.