Resolves #34 - Hook up TimesEditorVC to SecondVC
It was really just a matter of sticking some code to get the next view
controller into the editing part of the handler for SecondVC. I also
apparently forgot to give TimesEditorVC a storyboard ID so I had to do
that. I also apparently forgot to present the UIAlerC so I also had to
add some code to present it.