Resolved #35: add clear times functionality
I added the necessary workings in order to clear all of one's times,
including another alert to make sure you really want to do that.

In the process, I discovered that it won't update SecondViewController
unless you add Another time to the mix, so I need to fix that soon.