Resolved #39: test new fridrich cells
I've tested most of them and also added them to the OLL VC. I haven't
really done much with the PLL cells but I think that once I have the
code down for the OLL one the PLL one should be fine.

In fact, now that I think about it, I should have PLL subclass from OLL,
and just change the values of the pertinent fields.