Resolves #47, Resolves #48
Issue #47 - piece PLLVC together
I had to refactor code in order to put the PLLVC together. In fact, most
of this work, though not all, was done by resolving #48. I just had
to override a few methods.

Issue #48 - refactor OLLVC code
I had to refactor the code in order to reuse the class in when using the
PLLVC class. Otherwise, I would have had to rewrite this class entirely.

So now the PLL page works just as well as the OLL page, which is not to
say that improvements cannot be made.