Resolved #49: get all of the pll algs to fit
Well, at least resolved it mostly. It would be difficult, highly
difficult to get all of them to fit, so I'm reluctantly admitting to
defeat for now so that I can get the rest of this app running. For
example, I want to be able to sort the many many many cases between
known and unknown, so I want to get that sort of mechanism in place.