Fixes #55, Resolves #56
Issue #55 - Implement segue for Session VC
Issue #56 - Implement button for Session VC
Apparently these issues were one in the same: in order to have the back
button, I had to push the times VC on top of the Session VC, so that was
interesting. All I had to do to get TimesVC on top of SessionVC was to
push TimesVC without animation on top of SessionVC once the view loaded
for SessionVC.
*I'm calling SecondVC TimesVC here because it's shorter and more
descriptive.