Resolves #57: list sessions in SessionVC
Literally just have it listing the sessions. No more, no less. I made a
special SessionCell just for this purpose, but right now it's more like
a shell. Will probably need to make a special Session object that
perhaps holds all of the times, and its name, the start date, and the
end date. If the SessionCell could have a method that takes the Session
object and then assigns itself the right information, that would
probably be the best I could ask for.