Resolved #60 - create slider cell
I made a SliderCell class. In addition to the normal attributes of
nameLabel, it also has valueLabel and cellSlider. cellSlider is a
UISlider that can be used to set some value in settings. In this case,
it is 0.5 by default. Setting the user's personal default will be a job
for the SettingsVC.

I have not tested the SliderCell yet but it should update on its own.