Working on #52, Resolved #61
Issue #61 - Implement slider cell
I've implemented it, and I changed Storyboard to reflect that the last
view controller is supposed to be a SettingsViewController, but whenever
the value of the slider changes, the value on the label does not change,
which is concerning.

I'll probably have to fiddle with this some more.