Resolves #66 - timerVC gets slider val
timerVC (aka FirstVC) gets the slider value during viewWillAppear so
that it can change the minimum press duration of the longPress gesture
recognizer. timerVC also sets the slider value in the userdefs
(userdefaults) if the value in there doesn't make any sense.

I might trim the range of the slider's values but in any case this is
good.