Fixes #64 - PLLVC crashes again
The cause was super.viewdidload. OLLVC was doing things that shouldn't
be done to PLLVC, so I just stuck all of the unwanted stuff into a
function that OLLVC calls, then overwrote that function in PLLVC, so
that when super.viewDidLoad is called, it just calls that viewDidLoad
function that we get from UITableVC and nothing extraneous is done.