Fixed #36, Fixed #70
Issue #36 - Refresh SecondVC after clearing times
The TimesVC (formerly SecondVC) was not reflecting that all of the times
had been deleted when all of the times were cleared. I had to fix this
by changing around some code and boolean tests, as well as adding more
boolean tests.

Issue #70 - fix saving data bug
The data wasn't being saved after a deletion or a clear, so I had to add
that in. I'm still not totally sure if it's being saved after a clear
though.

I had to change both code in performInsertion and viewWillAppear for
TimesVC in order to fix this.