Fixed #71 - fix panning gestures
All sorts of gestures (long press, tap) were being mistaken for panning
gestures, so I had to fix that. I did this by mandating that the panning
gestures had to cross at least 25% of the screen in order to be
considered a panning gesture. I also had to make the function
absoluteValue because the abs() function that comes with the Swift
foundation seems to only return integers. That way I could tell whether
it was a large enough translation despite whether it was going left or
right.