Resolved #68, Working on #69
Issue #68 - create parent UITableVC
This was an easy one to solve - just put together all of the things that
I keep having to redo every single time I start coding one of these
things: have a titles array for your header titles, have the footer and
headers be the same color, have a footer for the entire table, change
the background of the entire tableview, deselect the rows when they're
selected, etc.

Issue #69 - implement parent UITableVC
I was working on this when I decided to commit. I've applied it (mostly)
in SessionVC so far, just as a test run. It seems to work very well. I
want to implement it on the non-complicated VCs first before moving on
to more complicated cases, like TimesVC or PLL/OLLVC.