Fixes #72, Resolves #69
Issue #72 - fix saving/getting of PLL known cases
I noticed offhand that the saving-getting of PLL known cases wasn't
sticking after every instance of the app running, so I figured out it
was because it would save known then get known, instead of getting then
saving. So I fixed that, and it should work now.

Issue #69 - implement parent UITableVC
My parent UITableVC is called TableVC, and it is essentially
streamlining my approach to TableVCs. I haven't implemented it across
all of my VCs because I think that would be too messy, but I have used
it in both SessionViewController and SettingsViewController. I'll
probably also implement it in the TimesEditorVC but I'll save that for a
different commit - I haven't even implemented times editing in that VC
yet.