Resolved #44 - implement stats
I've completely implemented stats. There were a few problems along the
way having to do with indices and array slices, but I resolved them. I
mostly had to make sure that TimesVC was using all of the resources that
I had for it. I might refactor the stats code to make it more
reproducible and modular.