Resolved #73 - refactor stats code
I wanted to try using arrays/dictionaries in order to generalize and
modularize and shorten and abstract some of my stats code so that I
could have the stats code become shorter and better and perhaps more
understandable. This is mostly pertaining to the function
availableStats. I also ended up making the init function finally
functional.