Resolves #81 - form cell class
I've added functions to make it easier for me to deal with things - such
as adding constraints. I have not implemented most of it yet, but to
start off, I just made a basic one that has its name but also has a
variable for a UITextField so that I can add it in the future.