Fixes #84 - fix stats updating
The statistics weren't updating, so what I had to do was call
tableView.reloadData() after updating the stats' data.

Then it turned out that there was a problem with it being 100, i.e. it
would have a negative start index, so I had to fix that as well using
max(calculated index, 0) so if the calculated version was negative, it
would just become 0.