Resolves #78 - create session class; fixes typo
Issue #78 - create session class
Created the session class and added functions that I thought would be
helpful/useful:
 - save
 - get
 - hasSavedBefore
 - append (a time value in milliseconds)
 - remove (a time value at an index)
 - remove (a time value at an index represented by an index path)
 - time (return the time value at an index if it is a correct index,
   otherwise return -1)
This should help me to create a better interface for sessions.

I also fixed a typo in StatisticsCalc. There was a q or a w after the
use of a max() function. I had to delete it because it was extraneous.