Resolves #75, Resolves #76
Issue #75 - add VC for adding session
Issue #76 - create VC class for adding a session
These two are really one in the same so I'm just putting them in the
same commit.
Made the VC class for adding a session, AddSessionViewController, whose
file is named AddSessionVC. I've been using VC as an abbreviation of
ViewController more and more over time, and I've found that it is
difficult to see the class I'm trying to find unless I shorten the name.
I can see the name AddSession so much more clearly if I shorten the
name. Eventually, I will rename all of the VCs similarly.
I copied and pasted code from SessionViewController into AddSessionVC
because there was code for a UITextFieldDelegate in SessionVC. I also
commented out code that I was using to test my form cell in SessionVC.
When I added the VC for the main storyboard, I added a segue, a
storyboard id, etc. I still haven't added the necessary code in
SessionVC to actually effect the segue, but everything is in there to
effect the changes.