Resolves #77 - integrate sessions VC
Issue #77 - integrate new VC class/storyboard with SessionVC
Tied opening up of AddSessionVC to SessionsVC "Add new..." button. Once
you're done with the label and click enter, you go back to SessionVC and
the new session is added.

I think I've fixed all of the problems with saving sessions as well.
There was a good fifteen minutes or so where I could not get it to stop
having array index out of bounds errors, but it was because I was
saving and getting the stored copy of arrays and dictionaries in the
wrong order; it should be saved right after one adds new material.