Fixes #86 - support choosing of sessions
This seemed like the logical next choice for this repository. SessionVC
now supports choosing sessions through a checkmark that is placed next
to its name in the SessionCell. Therefore, I also had to change
SessionCell in order for this to work properly, so SessionCell now also
has a variable called checkImage that will hold the checkmark image if
it is indeed the one to be checked off. I also added two helper
functions that either turn the SessionCell into one that has a checkmark
or one that does not. The image is blue, but I determine that
programmatically by changing the tint color of the image view.
I also have to make it a template from the start; however, I will do
this later because I'm currently just glad that it works.