Fixed #90, Encountered #89
Issue #90 - add deletion to session cells
I added deletion by swiping to the left which brings up a little delete
button. I then added a little code that deletes it and makes the
necessary changes to make it better.

Issue #89 - encountered bugs because of the saving of multiple sessions
I believe it was an optional that was unwrapped but there was a nil
inside. I had it save the session after deleting things and this seems
to work better.