Resolved #54 - be able to rearrange cases
I implemented #54 in both PLL and OLL. This commit only has the changes
for PLL though.

In doing this, I accidentally made it impossible to switch cases between
known and unknown. I might have to do something, i.e. make an
alert/actionsheet or something to be able to choose, or make some kind
of function that facilitates moving between the two sets of arrays. It
is annoying for either of them but it is especially so for PLLVC.