Fixed #94 - fix saving of sessions; index errors
Issue #94 - fix saving of sessions in infosharingTBC
I got it so that it saves not only the session properly (it was loadData
that was at fault for making everything screwy), it also saves
currentSession properly now too.

Index errors
StatisticsCalc was throwing errors about negative indices so I fixed
them by fixing some of my math.