Fixes #97, Working on #9
Issue #97 - adapt app for landscape
The timerVC did not work out so well in landscape mode so I've changed
the constraints in order to better adapt it.

Issue #9 - Transfer of times to computer
Worked a little on this issue, added a line to the plist that basically
allows for file sharing between iTunes and my app. I'm not sure how to
test this though because I don't know how I'm going to get iTunes to
recognize my app.