Resolved #99 - adjust settingsVC for more settings
I sought to adapt settingsVC to accommodate more settings, and in the
process, also created a new class, SettingsCell, that will replace
SliderCell. SettingsCell has an enum, SettingsCellType, that determines
what it looks like. Maybe I should change its name to style? Anyway,
there are two types: rerouter and slider. slider is for the particular
cell changes a float value, and rerouter is for going to different
pages.