Resolves #102 - add collection view for dates
Issue #102 - add collection view for displaying dates
More specifically, it determines how dates are displayed. I had to add a
special dateFormatCell in order to do this. I also had to learn some
things about the UICollectionViewController in order to get this in.

Working on this issue meant working on Issue #101, but I still have to
implement different sections of the UICollectionView in order to get
that working.