Closes #104 - create app icon
I created the app icon and added all of the necessary (and some
unnecessary) assets to the project.

I ended up creating the app icon by 3D modeling a Rubik's cube in
Blender and then taking a screenshot of the render, which I then opened
up in Photoshop, cleaned up a little, and then exported it to all of the
necessary sizes. I think in the future I should create an ImageMagick
script that will do that for me - well, it will be a Bash script that
uses ImageMagick tools. I think I can do that, anyway...