Fixes #105 - fix the separators for oll/pll
This was a hard one to crack - I couldn't figure out how to add an
image, the default imageView that comes with the UITableViewCell class,
in a custom layout without breaking the separators. I tried nearly
everything I think.

That was, until I realized that I could just make my own image view
instead of using the default one, and this solved all of the problems,
because then I could simply just remove the image view from its
superview. Problem solved, now all of the layout and everything works
again.