Fixed #112 - woops fix cubecasecell constraints
I messed up the constraints while playing around with the layout, and
messed everything up, but now it should be fine again.

Unfortunately, this means that I couldn't get everything nice and neat
into the content view, so that you could actually edit the cell
properly, but that's okay for now.