Resolved #111 - add button in add session...
Issue #111 - add button in add session VC that will trigger adding a new
session of name [insert today's date here]
I like assigning the names of stuff to the date so I know when it was
that I did that thing, so this should make it easy for me to do that. I
think I might make a setting that will let you toggle between having it
be a date or not, or perhaps you can have it auto-populate the date
without you having to press a button, etc.

I still need to fix the layout of the button for the date vs the
UITextField, but I'm saving that for a different issue.