Fixed #117 - fix positioning of cases w 1 alg
Issue #117 - fix positioning of cases with only one alg
The algorithm would either be in the center, too close to the top, or
too close to the bottom. I aimed to fix that here. A few things were
happening that made this a problem

1. setupViews() was being called before anything was ever changed from
its default values, so all of the labels had a "Sample Item," which
ruined the logic in setupViews()
2. If you call setupViews() without changing the default values, you're
essentially just making a bunch of cells that think they have two cases,
thus a lot of the cells just had algLabel inserted into the top version,
resulting in a cell with a single algorithm that is positioned too close
to the top.
3. If you call setupViews() twice without removing any of the views that
may need to have their constraints altered, you'll have two constraints
for one view that cannot be satisfied simultaneously, resulting in
strange/unexpected behavior.
4. Cells get reused, so sometimes they would have a double alg cell, but
then would suddenly only have one alg, and these constraints would
conflict with each other, resulting in unsatisfiable constraints.

I'm glad I got this sorted out because all of the unsatisfiable
constraints warnings were rather verbose and made it difficult to see
any of the behavior of the actual app.