Shortened a lot of the names of swift files
I think "ViewController" is the most-often used phrase in the names of
the class files so I just had to change their file names to [insert
descriptor here]VC from [insert descriptor here]ViewController.
"ViewController" was often much longer than any descriptor that I'd
given them, resulting in an abundance of kind of useless information,
which made it difficult to find the file I was looking for.