Resolves #119 - update icon - black background
I updated the icon by making the background much darker than it used to
be. I hope this will make the colors pop more.

I also got the chance to try out a script that I wrote to make the ios
assets for me, because I got tired of making them myself.