Changed some of the settings
Now it should be building with my own developer account instead of some
random developer account, and is using the bundle ID that I made for
Cube Timer.

Apparently some of the assets were also changed (mostly the
contents.jsons, for what reason I do not know), but this was probably
something xcode did on its own.