Deleted old code/comments from CubeCaseCell
If I'm going to work on the positioning of elements when editing a case
cell, I want to sort of start off with a fresh slate, so to speak.

There were so many old comments, I could barely tell what was going on
in the code, so I'm probably going to be just refactoring for the next
few commits, including this one.