Added equalityContsraint function
Another sort-of wrapper function, but also because I tend to have very
few patterns that I want to follow other than I want item_a's center y
point to be equal to item_x's center y point with a difference/constant
of 0 and a multiplier of 1. So really, all I'd need to fill in is item_a
and item_b. I left the NSLayoutAttribute also as a parameter just
because I think it might be useful for testing.