Added EditableCubeCaseCell
Subclassed EditableCubeCaseCell from CubeCaseCell because I didn't want
to ruin CubeCaseCell, and I would have had to repeat a lot of the same
code anyway. I put in the initializers and I started fleshing out its
setupViews function.

I also added an entry for EditableCubeCaseCell to the
InfoSharingTabController so that when the time comes, it will be
relatively simple to get its cell reuse identifier.