Fixed #92 - fix positioning of elements when edit
Issue #92 - fix positioning of elements when editing a case cell
I finally fixed this issue. I think it was a combination of not adding
the constraints to the contentView and not adding the subviews to the
contentView in the init. I could verify what exactly makes the
difference I suppose and report back in the next commit.

I'm also just super glad that it finally works.