What fixed it was contentview's constraints
I think the way I did the constraints this time was consistent enough
because I used a function that the constraints finally worked.
Commenting out setting the contentView's bounds or the placement of
adding the subviews to the contentView did nothing. So I believe that it
was really just the constraints.